refactor(analysis): split api surface symbol scanners
